rcon-telegram-bot/requirements.txt
aiogram>=3.4,<4.0
python-dotenv>=1.0,<2.0
PyYAML>=6.0,<7.0
mcrcon>=0.7,<1.0
